Augment the fix for HungRenderDialogView::ButtonPressed.
We need to check GetRenderProcessHost() (also fixed the typo in the if statement).

BUG=http://crbug.com/52293
TEST=None

Review URL: http://codereview.chromium.org/3348009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@58492 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f61739a5655dfc16f653a027f2ba1567d04a33cb

+++ chrome/browser/views/hung_renderer_view.cc
@@ -318,7 +318,7 @@ views::View* HungRendererDialogView::GetContentsView() {
 void HungRendererDialogView::ButtonPressed(
     views::Button* sender, const views::Event& event) {
   if (sender == kill_button_) {
-    if (!contents_) {
+    if (contents_ && contents_->GetRenderProcessHost()) {
       // Kill the process.
       TerminateProcess(contents_->GetRenderProcessHost()->GetHandle(),
                        ResultCodes::HUNG);
